weather_update/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
meteostat==1.6.8
numpy==2.0.2
protobuf==5.29.3
geopy==2.4.1